Some changes to schedule
