+++ frontend/components/calendar/colors.ts
@@ -1,11 +1,11 @@
 
 const colors : string[] = [
-    "#6A6B83",
-    "#96C3CE",
-    "#0FA3B1",
-    "#78B85A",
-    "#6F5060",
-    "#FA9F42",
+    "#66689d",
+    "#609db0",
+    "#609d71",
+    "#9f5454",
+    "#a85282",
+    "#a2764c",
 ]
 
 export {colors}
